requirements.txt
Flask==3.0.0
beautifulsoup4==4.12.2
requests==2.31.0
networkx==3.2.1
python-dateutil==2.8.2
lxml==4.9.3
gunicorn==21.2.0